refactor : service package 대신 domain 객체 생성
