+++ src/main/java/lotto/Application.java
@@ -3,7 +3,6 @@ package lotto;
 import lotto.controller.LottoController;
 import lotto.handler.InputHandler;
 import lotto.handler.OutputHandler;
-import lotto.service.LottoService;
 import lotto.view.ConsoleInput;
 import lotto.view.ConsoleOutput;
 
@@ -13,8 +12,7 @@ public class Application {
 
         final InputHandler inputHandler = new ConsoleInput();
         final OutputHandler outputHandler = new ConsoleOutput();
-        final LottoService lottoService = new LottoService();
 
-        new LottoController(inputHandler, outputHandler, lottoService).run();
+        new LottoController(inputHandler, outputHandler).run();
     }
 }
