+++ src/main/java/lotto/domain/GenerateLotto.java
@@ -1,4 +1,4 @@
-package lotto.service;
+package lotto.domain;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import lotto.domain.wrapper.Lotto;
@@ -10,9 +10,19 @@ import java.util.stream.IntStream;
 
 import static lotto.handler.ConstantsHandler.*;
 
-public class LottoService {
+public class GenerateLotto {
 
-    public List<Lotto> generateBuyLottos(Money money) {
+    private final List<Lotto> buyLottos;
+
+    private GenerateLotto(Money money) {
+        this.buyLottos = generateBuyLottos(money);
+    }
+
+    public static GenerateLotto create(Money money) {
+        return new GenerateLotto(money);
+    }
+
+    private List<Lotto> generateBuyLottos(Money money) {
         int ticket = money.getTicket();
 
         List<Lotto> buyLottos = IntStream.range(0, ticket)
@@ -20,5 +30,9 @@ public class LottoService {
                 .collect(Collectors.toList());
 
         return buyLottos;
+    }
+
+    public List<Lotto> getBuyLottos() {
+        return List.copyOf(buyLottos);
     }
 }
